requirements.txt
python-dotenv==1.1.1
openai==1.105.0
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
PyGithub==2.8.1
Flask==3.1.2
nbformat
hawk-python-sdk
email